pyproject.toml
[project]
name = "marketplace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aioredis>=2.0.1",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = []
